add quick + dirty text rendering test
Note
----
    Rendering Text like this is most likely not a good idea, we create a
separate image for each string of text, which could lead to a lot of
created textures if you are using any dynamic/changing text. I am going
to attempt to build a better version using a font texture atlas and
batch rendering to produce a more "real-world" and performant text
rendering strategy.